fix(agent-panel): use logo for create pull request
Summary:
- Adds a template logo asset for the create pull request action.
- Uses the logo only when the pull request status opens a create URL.
- Keeps existing pull request states on the pull request glyph.

Rationale:
- The URL-backed create row is an external action, while open/draft/merged
  states still represent pull request status.

Tests:
- xcodebuild test -workspace apps/mac/supaterm.xcworkspace -scheme supaterm -destination "platform=macOS" -only-testing:supatermTests/IconAssetTests CODE_SIGNING_ALLOWED=NO CODE_SIGNING_REQUIRED=NO CODE_SIGN_IDENTITY="" -skipMacroValidation
